# util/deephash: improve cycle detection