@josharian 2021-07-20
depaware is https://github.com/tailscale/depaware/. Install, run `depaware -update ./cmd/tailscale{,d}`.

The code LGTM, but I'd like to get resolution on using unsafe.Pointer (and ideally start using it) before submitting.